requirements.txt
python-telegram-bot==8.0
psycopg2==2.7.3.1
requests==2.18.4